StreamlitClassifierDemo/requirements.txt
pillow~=11.1.0
streamlit~=1.44.1
torch~=2.5.1
numpy~=2.0.2
tensorflow~=2.18.0
torchvision~=0.20.1
timm~=1.0.14
pandas~=2.2.3
altair~=5.5.0
pyarrow~=14.0.2